Run full scan with iw to avoid errors in iwlist
iwlist can cause an "Argument list too long" error if there
are too many networks nearby. See:

https://stackoverflow.com/questions/42623942/failed-to-read-scan-data-argument-list-too-long

On the other hand, iw doesn't seem to support the "quick-scan"
feature (returning the cached list of scanned interfaces)
that iwlist provides, so we continue to use both for now.

We could possibly drop the quick-scan feature from the backend
and cache manually in the frontend; that would allow us to
drop the iwlist dependency, but it would also require changes
to the AwesomeWM applet (gobo-awesome-gobonet).